04_nyc_taxi_trips/requirements.txt

apache-airflow-providers-amazon
apache-airflow-providers-postgres
